pyproject.toml
[project]
name = "expense-manager-cronjob"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "psycopg2-binary>=2.9.10",
]
